scripts/pyproject.toml
[project]
name = "scripts"
version = "0.1.0"
description = "Helper scripts"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=12.1.0",
]
